% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiucthemes.R
\docType{package}
\name{idahothemes-package}
\alias{idahothemes}
\alias{idahothemes-package}
\title{idahothemes: 'R' 'Markdown' Themes for 'UofI' Documents and Presentations}
\description{
A set of custom 'R' 'Markdown' templates for documents and
   presentations with the University of Idaho
   color scheme and identity standards.
}
\details{
Trial run on RMarkdown templates
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aatrujillob/idahothemes}
}

}



